[RA] Fixes crash on 64bit linux builds.
Fixes crash when a bib is removed.
Alters TD code to be the same for consistency.

+++ tiberiandawn/defines.h
@@ -1639,7 +1639,7 @@ typedef enum RadioMessageType : unsigned char
 **	with cell resolution. The COORD type is used for map coordinates that
 **	have a lepton resolution.
 */
-typedef unsigned long COORDINATE;
+typedef unsigned COORDINATE;
 typedef signed short CELL;
 
 typedef unsigned short TARGET;
